requirements.txt
aiofiles==22.1.0
aiogram==3.0.0b6
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.18.0
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
crc16==0.1.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
munch==2.5.0
numpy==1.24.1
orjson==3.8.4
pycparser==2.21
pydantic==1.10.4
PyNaCl==1.5.0
python-dotenv==0.21.0
python-jsonrpc==0.10.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
ton==0.26
tonsdk==1.0.9
typing_extensions==4.4.0
urllib3==1.26.14
yarl==1.8.2